refactor bitflags!
