+++ macros/src/cargo.rs
@@ -1,0 +1,5 @@
+use std::env;
+
+pub(crate) fn package_name() -> String {
+    env::var("CARGO_PKG_NAME").unwrap_or_else(|_| "<unknown>".to_string())
+}
