+++ macros/src/items.rs
@@ -1,3 +1,4 @@
 //! Procedural macros that expand to items
 
+pub(crate) mod bitflags;
 pub(crate) mod timestamp;
